agent/requirements.txt
# Sake Sensei Agent - Dependencies

# Amazon Bedrock AgentCore SDK
bedrock-agentcore-runtime>=0.1.0

# Strands Agents Framework
strands-agents>=0.1.0

# MCP (Model Context Protocol) Client
mcp>=0.1.0

# Environment management
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0